requirements.txt
flask
python-dotenv
flask_pymongo
flask-smorest
sqlalchemy
flask-sqlalchemy